style: 견적 요청 레이아웃 수정
