+++ src/app/(with-protected)/request/layout.tsx
@@ -1,22 +1,28 @@
-import DefaultLayout from "@/components/layout/DefaultLayout"
-import ExampleLayout from "@/components/layout/ExampleLayout"
-import Header from "@/components/layout/Header"
+"use client";
 
-export default function Layout({children} : {children : React.ReactNode}) {
-    return(
-        <div
-            className="
-                min-h-screen
-                bg-bg-200   
-            "
-        >
-            <Header>
-                {/* 여기에 서브 헤더 컴포넌트 추가 */}
-                <ExampleLayout />
-            </Header>            
-            <DefaultLayout>
-                {children}
-            </DefaultLayout>
+import DefaultLayout from "@/components/layout/DefaultLayout";
+import Header from "@/components/layout/Header";
+import ProgressBar from "./_components/ProgressBar";
+import { useEffect, useState } from "react";
+import { useFormWizard } from "@/context/FormWizardContext";
+
+export default function Layout({ children }: { children: React.ReactNode }) {
+  const { currentStep } = useFormWizard();
+  const [progress, setProgress] = useState(0);
+
+  useEffect(() => {
+    setProgress((currentStep + 1) * 25);
+  }, [currentStep]);
+
+  return (
+    <div className=" min-h-screen bg-bg-200">
+      <Header>
+        <div className="w-full py-4 space-y-4 lg:py-8 lg:space-y-6">
+          <h1 className="text-lg font-semibold lg:text-2xl">견적 요청</h1>
+          {currentStep !== 4 && <ProgressBar progress={progress} />}
         </div>
-    )
+      </Header>
+      <DefaultLayout>{children}</DefaultLayout>
+    </div>
+  );
 }
