+++ src/app/layout.tsx
@@ -16,9 +16,7 @@ export default function RootLayout({
   return (
     <html lang="ko">
       <body className="min-h-full h-full">
-        <main>
-          <Providers>{children}</Providers>
-        </main>
+        <Providers>{children}</Providers>
       </body>
     </html>
   );
