--- src/types/index.ts
@@ -1,3 +1,0 @@
-export * from "./support.types";
-export * from "./chip.types";
-export * from "./request.types";
